[new-style] (squash) split out optimizer module and port most of server to use new shape types and optimizer. still need JSON instances for TagExpr (the actual Value type later) and to fix dragUpdate. also deprecate mainretfinal
