+++ src/Server.hs
@@ -3,11 +3,12 @@
 {-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE DeriveGeneric #-}
 {-# LANGUAGE DuplicateRecordFields #-}
+{-# LANGUAGE RankNTypes, NoMonomorphismRestriction #-}
 
 module Server where
 import Shapes
 import Computation
-import Utils (Autofloat)
+import Utils (Autofloat, r2f)
 import GHC.Generics
 import Data.Monoid (mappend)
 import Data.Text (Text)
@@ -21,10 +22,13 @@ import Data.Maybe (fromMaybe)
 import GHC.Float (float2Double)
 import Network.WebSockets.Connection
 import System.Time
+import System.Random
 import Debug.Trace
+import qualified NewStyle as NS
+import qualified Optimizer as O
+import qualified Data.Map as M
 import qualified Data.Text                 as T
 import qualified Data.Text.IO              as T
-import qualified Runtime                   as R
 import qualified Network.WebSockets        as WS
 import qualified Network.Socket            as S
 import qualified Network.WebSockets.Stream as Stream
@@ -32,27 +36,57 @@ import qualified Control.Exception         as Exc (catch, ErrorCall)
 
 
 -- Types used by the server, mainly for translation to JSON
-type ServerState = R.State
-data FeedBack = Cmd Command | Drag DragEvent | Update UpdateShapes deriving (Show, Generic)
-data Command = Command { command :: String } deriving (Show, Generic)
-data DragEvent = DragEvent { name :: String, xm :: Float, ym :: Float } deriving (Show, Generic)
-data UpdateShapes = UpdateShapes { objs :: [Obj] } deriving (Show, Generic)
-data Frame = Frame { flag :: String, objs :: [Obj] } deriving (Show, Generic)
-instance FromJSON FeedBack
+type ServerState = NS.RState
+
+data Feedback = Cmd Command 
+                | Drag DragEvent 
+                | Update UpdateShapes 
+     deriving (Generic)
+
+data Command = Command { command :: String } 
+     deriving (Show, Generic)
+
+data DragEvent = DragEvent { name :: String, 
+                             xm :: Float, 
+                             ym :: Float } 
+     deriving (Show, Generic)
+
+data UpdateShapes = UpdateShapes { 
+                       shapes :: [NS.Shape Double] 
+                    } deriving (Show, Generic)
+
+data Frame = Frame { flag :: String, 
+                     shapes :: [NS.Shape Double] 
+                   } deriving (Show, Generic)
+
+-- TODO write these. do we need FromJSON instance for "a"? or should we just convert to floats?
+instance FromJSON (NS.TagExpr a) where
+         parseJSON = error "TODO fromjson tagexpr"
+
+instance ToJSON (NS.TagExpr a) where
+         toJSON x = error "TODO tojson tagexpr"
+
+instance FromJSON Feedback
 instance FromJSON Command
 instance FromJSON DragEvent
 instance FromJSON UpdateShapes
 instance ToJSON Frame
 
+wsSendJSON :: WS.Connection -> (NS.Shape Double) -> IO ()
+wsSendJSON conn shape = WS.sendTextData conn $ encode shape
 
-wsSendJSON :: ToJSON j => WS.Connection -> j -> IO ()
-wsSendJSON conn obj = WS.sendTextData conn $ encode obj
+-- TODO use the more generic wsSendJSON?
+wsSendJSONList :: WS.Connection -> ([NS.Shape Double]) -> IO ()
+wsSendJSONList conn shapes = WS.sendTextData conn $ encode shapes
+
+wsSendJSONFrame :: WS.Connection -> Frame -> IO ()
+wsSendJSONFrame conn frame = WS.sendTextData conn $ encode frame
 
 -- | 'servePenrose' is the top-level function that "Main" uses to start serving
 --   the Penrose Runtime.
 servePenrose :: String  -- the domain of the server
              -> Int  -- port number of the server
-             -> R.State  -- initial state of Penrose Runtime
+             -> NS.RState  -- initial state of Penrose Runtime
              -> IO ()
 servePenrose domain port initState = do
      putStrLn "Starting Server..."
@@ -94,28 +128,21 @@ application :: ServerState -> WS.ServerApp
 application s pending = do
     conn <- WS.acceptRequest pending
     WS.forkPingThread conn 30 -- To keep the connection alive
-    wsSendJSON conn (R.objs s)
-    loop conn (step s)
+    wsSendJSONList conn (NS.shapesr s)
+    loop conn (O.step s)
 
--- Apply computations N times post-optimization (TODO: just a terrible hack until explicit comp graph is built)
-computeN :: (Autofloat a) => Int -> [Obj] -> [R.ObjComp a] -> [Obj]
-computeN n objs comps = let res = iterate (flip R.computeOnObjs_noGrad comps) objs in
-                        res !! n -- hopefully doesn't use too much space
-
-loop :: WS.Connection -> R.State -> IO ()
+loop :: WS.Connection -> NS.RState -> IO ()
 loop conn s
-    | R.optStatus (R.params s) == R.EPConverged = do
+    | NS.optStatus (NS.paramsr s) == NS.EPConverged = do
         putStrLn "Optimization completed."
-        putStrLn ("Current weight: " ++ (show $ R.weight (R.params s)))
-        let objsComputed = computeN 5 (R.objs s) (R.comps s)
-        putStrLn "Final computations applied"
-        -- putStrLn $ "Final objs:\n" ++ show objsComputed
-        wsSendJSON conn Frame { flag = "final", objs = objsComputed }
+        putStrLn ("Current weight: " ++ (show $ NS.weight (NS.paramsr s)))
+        wsSendJSONFrame conn (Frame { flag = "final", 
+                                shapes = (NS.shapesr s) :: ([NS.Shape Double]) })
         processCommand conn s
-    | R.autostep s = stepAndSend conn s
+    | NS.autostep s = stepAndSend conn s
     | otherwise = processCommand conn s
 
-processCommand :: WS.Connection -> R.State -> IO ()
+processCommand :: WS.Connection -> NS.RState -> IO ()
 processCommand conn s = do
     --putStrLn "Receiving Commands"
     msg_json <- WS.receiveData conn
@@ -123,45 +150,63 @@ processCommand conn s = do
     case decode msg_json of
         Just e -> case e of
             Cmd (Command cmd)  -> executeCommand cmd conn s
-            Drag (DragEvent name xm ym)  -> dragUpdate name xm ym conn s
-            Update (UpdateShapes objs)  -> updateShapes objs conn s
+            Drag (DragEvent name xm ym)  -> error "TODO: IMPLEMENT DRAG UPDATE" -- dragUpdate name xm ym conn s
+            Update (UpdateShapes shapes)  -> updateShapes shapes conn s
         Nothing -> error "Error reading JSON"
 
-updateShapes newObjs conn s = if R.autostep s then stepAndSend conn news else loop conn news
-    where
-        news = s { R.objs = newObjs,
-                   R.params = (R.params s) { R.weight = R.initWeight, R.optStatus = R.NewIter }}
+toPolymorphics :: [NS.Shape Double] -> (forall a . (Autofloat a) => [NS.Shape a])
+toPolymorphics = map toPolymorphic
+   where toPolymorphic :: NS.Shape Double -> (forall a . (Autofloat a) => NS.Shape a)
+         toPolymorphic (ctor, properties) = (ctor, M.map toPolyProperty properties)
 
-dragUpdate :: String -> Float -> Float -> WS.Connection -> R.State -> IO ()
-dragUpdate name xm ym conn s = if R.autostep s then stepAndSend conn news else loop conn news
-    where
-        newObjs = map (\x ->
-                if getName x == name
-                    then setX (xm + getX x) $ setY (-ym + getY x) x
-                    else x)
-            (R.objs s)
-        news = s { R.objs = newObjs,
-                   R.params = (R.params s) { R.weight = R.initWeight, R.optStatus = R.NewIter }}
+         toPolyProperty :: NS.TagExpr Double -> (forall a . (Autofloat a) => NS.TagExpr a)
+         toPolyProperty e@(NS.OptEval v) = NS.OptEval v
+         toPolyProperty e@(NS.Done v) =
+               case v of
+               TNum n -> NS.Done $ TNum $ r2f n
+               -- Not sure why these have to be rewritten from scratch...
+               TBool x -> NS.Done $ TBool x
+               TStr x ->  NS.Done $ TStr x
+               TInt x ->  NS.Done $ TInt x
+               _ -> error "finish double -> polymorphic value conversion" -- TODO finish
+               -- TList x ->  NS.Done $ TList x
+               -- TPt x ->  NS.Done $ TPt x
 
-executeCommand :: String -> WS.Connection -> R.State -> IO ()
+updateShapes :: ([NS.Shape Double]) -> Connection -> NS.RState -> IO ()
+updateShapes newShapes conn s = if NS.autostep s then stepAndSend conn news else loop conn news
+    where
+        news = s { NS.shapesr = toPolymorphics newShapes,
+                   NS.paramsr = (NS.paramsr s) { NS.weight = NS.initWeight, NS.optStatus = NS.NewIter }}
+
+-- TODO implement this
+-- dragUpdate :: String -> Float -> Float -> WS.Connection -> NS.RState -> IO ()
+-- dragUpdate name xm ym conn s = if NS.autostep s then stepAndSend conn news else loop conn news
+--     where
+--         newShapes = map (\x ->
+--                 if getName x == name
+--                     then setX (xm + getX x) $ setY (-ym + getY x) x
+--                     else x)
+--             (NS.shapesr s)
+--         news = s { NS.shapesr = newShapes,
+--                    NS.paramsr = (NS.paramsr s) { NS.weight = NS.initWeight, NS.optStatus = NS.NewIter }}
+
+executeCommand :: String -> WS.Connection -> NS.RState -> IO ()
 executeCommand cmd conn s
     | cmd == "resample" = resampleAndSend conn s
     | cmd == "step"     = stepAndSend conn s
-    | cmd == "autostep" = loop conn (s { R.autostep = not $ R.autostep s })
+    | cmd == "autostep" = loop conn (s { NS.autostep = not $ NS.autostep s })
     | otherwise         = putStrLn ("Can't recognize command " ++ cmd)
 
-resampleAndSend, stepAndSend :: WS.Connection -> R.State -> IO ()
+resampleAndSend, stepAndSend :: WS.Connection -> NS.RState -> IO ()
 resampleAndSend conn s = do
-    let (objs', rng') = R.sampleConstrainedState (R.rng s) (R.objs s) (R.constrs s)
-    let nexts = s { R.objs = objs', R.rng = rng',
-                    R.params = (R.params s) { R.weight = R.initWeight, R.optStatus = R.NewIter } }
-    wsSendJSON conn (R.objs nexts)
-    loop conn nexts
-stepAndSend conn s = do
-    let nexts = step s
-    wsSendJSON conn (R.objs nexts)
+    let shapes' = NS.shapesr s
+                    -- TODO: shapes', rng' = NS.sampleConstrainedState (NS.rng s) (NS.shapesr s) (NS.constrs s)
+    let nexts = s { NS.shapesr = shapes', -- NS.rng = rng',
+                    NS.paramsr = (NS.paramsr s) { NS.weight = NS.initWeight, NS.optStatus = NS.NewIter } }
+    wsSendJSONList conn shapes'
     loop conn nexts
 
-step :: R.State -> R.State
-step s = s { R.objs = objs', R.params = params' }
-        where (objs', params') = R.stepObjs (float2Double R.calcTimestep) (R.params s) (R.objs s)
+stepAndSend conn s = do
+    let nexts = O.step s
+    wsSendJSONList conn ((NS.shapesr nexts) :: ([NS.Shape Double]))
+    loop conn nexts
